Design : addDialog ui수정
ui통일, 이벤트알림 디테일로 이동
