+++ app/src/main/res/values/themes.xml
@@ -12,7 +12,7 @@
     <!-- CustomDialog 배경 자르기 -->
     <style name="CustomAlertDialog">
         <!-- 꽉 채운 화면에 배경 없이 (custom view 영역만 보일 것이다.) -->
-        <item name="android:windowBackground">@android:color/transparent</item>
+        <item name="android:windowBackground">@android:color/white</item>
         <item name="android:windowMinWidthMajor">100%</item>
         <item name="android:windowMinWidthMinor">100%</item>
